feat: massivly optimise assements overview with components
